YT-18038: All methods to Rich Constrained commit_hash:cb800ba895c7e16f1f99275f563878e33dd009a7
