+++ yt/yt/client/ypath/public.h
@@ -14,7 +14,9 @@ YT_DEFINE_ERROR_ENUM(
 
 ////////////////////////////////////////////////////////////////////////////////
 
-class TRichYPath;
+template <class... TValidator>
+class TConstrainedRichYPath;
+using TRichYPath = TConstrainedRichYPath<>;
 
 ////////////////////////////////////////////////////////////////////////////////
