+++ yt/yt/client/ypath/rich.h
@@ -1,226 +1,15 @@
 #pragma once
 
 #include "public.h"
-
-#include <yt/yt/client/chunk_client/read_limit.h>
-
-#include <yt/yt/client/security_client/public.h>
-
-#include <yt/yt/client/table_client/column_sort_schema.h>
-#include <yt/yt/client/table_client/schema.h>
-
-#include <yt/yt/client/transaction_client/public.h>
-
-#include <yt/yt/core/yson/public.h>
-
-#include <yt/yt/core/ytree/attributes.h>
-
-#include <yt/yt/core/compression/public.h>
-
-#include <yt/yt/library/erasure/public.h>
+#include "rich_constrained.h"
 
 namespace NYT::NYPath {
 
 ////////////////////////////////////////////////////////////////////////////////
 
 //! YPath string plus attributes.
-class TRichYPath
-{
-public:
-    TRichYPath() noexcept = default;
-
-    TRichYPath(const TRichYPath& other) noexcept;
-    TRichYPath& operator=(const TRichYPath& other) noexcept;
-
-    TRichYPath(TRichYPath&& other) noexcept = default;
-    TRichYPath& operator=(TRichYPath&& other) noexcept = default;
-
-    TRichYPath(const char* path);
-    TRichYPath(TYPath path);
-    TRichYPath(TYPath path, const NYTree::IAttributeDictionary& attributes);
-
-    static TRichYPath Parse(TStringBuf str);
-    TRichYPath Normalize() const;
-
-    const TYPath& GetPath() const;
-    void SetPath(TYPath path);
-
-    const NYTree::IAttributeDictionary& Attributes() const;
-    NYTree::IAttributeDictionary& Attributes();
-
-    void Save(TStreamSaveContext& context) const;
-    void Load(TStreamLoadContext& context);
-
-    // Attribute accessors.
-    // "append"
-    bool GetAppend(bool defaultValue = false) const;
-    void SetAppend(bool value);
-
-    // "teleport"
-    bool GetTeleport() const;
-
-    // "primary"
-    bool GetPrimary() const;
-
-    // "foreign"
-    bool GetForeign() const;
-    void SetForeign(bool value);
-
-    // "read_via_exec_node"
-    bool GetReadViaExecNode() const;
-    void SetReadViaExecNode(bool value);
-
-    // "columns"
-    std::optional<std::vector<std::string>> GetColumns() const;
-    void SetColumns(const std::vector<std::string>& columns);
-
-    // "rename_columns"
-    std::optional<NTableClient::TColumnRenameDescriptors> GetColumnRenameDescriptors() const;
-
-    // "ranges"
-    //! COMPAT(ignat): also "lower_limit" and "upper_limit"
-    //! This method returns legacy read ranges completely ignoring key bounds.
-    std::vector<NChunkClient::TLegacyReadRange> GetRanges() const;
-
-    //! Get ranges, possibly transforming legacy keys into new key bounds using provided comparator.
-    //! This method throws if comparator is not present and keys are present.
-    //! In case when at least one column is non-ascending, some requirements are even stronger,
-    //! refer to the implementation for details.
-    std::vector<NChunkClient::TReadRange> GetNewRanges(
-        const NTableClient::TComparator& comparator = NTableClient::TComparator(),
-        const NTableClient::TKeyColumnTypes& conversionTypeHints = {}) const;
-
-    //! Check whether ranges contain a range with row_index specified in either
-    //! limit. This method is intended to be more lightweight than #GetNewRanges
-    //! and it does not require a comparator.
-    bool HasRowIndexInRanges() const;
-
-    void SetRanges(const std::vector<NChunkClient::TReadRange>& ranges);
-    bool HasNontrivialRanges() const;
-
-    // "file_name"
-    std::optional<TString> GetFileName() const;
-
-    // "executable"
-    std::optional<bool> GetExecutable() const;
-
-    // "format"
-    NYson::TYsonString GetFormat() const;
-
-    // "schema"
-    NTableClient::TTableSchemaPtr GetSchema() const;
-
-    // "sorted_by"
-    NTableClient::TSortColumns GetSortedBy() const;
-    void SetSortedBy(const NTableClient::TSortColumns& value);
-
-    // "row_count_limit"
-    std::optional<i64> GetRowCountLimit() const;
-
-    // "timestamp"
-    std::optional<NTransactionClient::TTimestamp> GetTimestamp() const;
-
-    // "retention_timestamp"
-    std::optional<NTransactionClient::TTimestamp> GetRetentionTimestamp() const;
-
-    // "output_timestamp"
-    std::optional<NTransactionClient::TTimestamp> GetOutputTimestamp() const;
-
-    // "optimize_for"
-    std::optional<NTableClient::EOptimizeFor> GetOptimizeFor() const;
-
-    // "chunk_format"
-    std::optional<NChunkClient::EChunkFormat> GetChunkFormat() const;
-
-    // "compression_codec"
-    std::optional<NCompression::ECodec> GetCompressionCodec() const;
-
-    // "erasure_codec"
-    std::optional<NErasure::ECodec> GetErasureCodec() const;
-
-    // "auto_merge"
-    bool GetAutoMerge() const;
-
-    // "transaction_id"
-    std::optional<NObjectClient::TTransactionId> GetTransactionId() const;
-
-    // "security_tags"
-    std::optional<std::vector<NSecurityClient::TSecurityTag>> GetSecurityTags() const;
-
-    // "bypass_artifact_cache"
-    bool GetBypassArtifactCache() const;
-
-    // "schema_modification"
-    NTableClient::ETableSchemaModification GetSchemaModification() const;
-
-    // "partially_sorted"
-    bool GetPartiallySorted() const;
-
-    // "chunk_unique_keys"
-    std::optional<bool> GetChunkUniqueKeys() const;
-
-    // "copy_file"
-    std::optional<bool> GetCopyFile() const;
-
-    // "chunk_sort_columns"
-    std::optional<NTableClient::TSortColumns> GetChunkSortColumns() const;
-
-    // "cluster"
-    std::optional<std::string> GetCluster() const;
-    void SetCluster(const std::string& value);
-
-    // "clusters"
-    std::optional<std::vector<std::string>> GetClusters() const;
-    void SetClusters(const std::vector<std::string>& value);
-
-    // "create"
-    std::variant<bool, NYTree::IAttributeDictionaryPtr> GetCreate() const;
-
-    // "versioned_read_options"
-    NTableClient::TVersionedReadOptions GetVersionedReadOptions() const;
-
-    // "versioned_write_options"
-    NTableClient::TVersionedWriteOptions GetVersionedWriteOptions() const;
-
-    // "access_method"
-    std::optional<TString> GetAccessMethod() const;
-
-    // "input_query"
-    std::optional<TString> GetInputQuery() const;
-
-private:
-    TYPath Path_;
-    NYTree::IAttributeDictionaryPtr Attributes_;
-};
-
-bool operator==(const TRichYPath& lhs, const TRichYPath& rhs);
-
-////////////////////////////////////////////////////////////////////////////////
-
-void FormatValue(TStringBuilderBase* builder, const TRichYPath& path, TStringBuf spec);
-
-std::vector<TRichYPath> Normalize(const std::vector<TRichYPath>& paths);
-
-void Serialize(const TRichYPath& richPath, NYson::IYsonConsumer* consumer);
-void Deserialize(TRichYPath& richPath, NYTree::INodePtr node);
-void Deserialize(TRichYPath& richPath, NYson::TYsonPullParserCursor* cursor);
-
-void ToProto(TString* protoPath, const TRichYPath& path);
-void FromProto(TRichYPath* path, const TString& protoPath);
-
-void ToProto(std::string* protoPath, const TRichYPath& path);
-void FromProto(TRichYPath* path, const std::string& protoPath);
-
-////////////////////////////////////////////////////////////////////////////////
-
-const std::vector<TString>& GetWellKnownRichYPathAttributes();
+using TRichYPath = TConstrainedRichYPath<>;
 
 ////////////////////////////////////////////////////////////////////////////////
 
 } // namespace NYT::NYPath
-
-template <>
-struct THash<NYT::NYPath::TRichYPath>
-{
-    size_t operator()(const NYT::NYPath::TRichYPath& richYPath) const;
-};
